Add files via upload
These 3 files are the regression results. In rg_summary_all.txt, the returns of all 26 factors are regressed the return of each asset. In rg_summary_each.txt, the return of each factor is regressed on each asset's returns separately. In the rg_summary_prtl.txt, the returns of each factor are regressed on the return of an equally weighted portfolio consisting of the 5 assets.